Refactor challenge controllers
