+++ api/src/controllers/challengesControllers.js
@@ -74,7 +74,25 @@ const getAchieved = async(req, res) => {
     }
 }
 
+const getCurrent = async(req, res) => {
+    try {
+        const { user_id} = req.body;
+
+        const user = await getConnection().get("users").find({ user_id: user_id }).value();
+        if (!user) return res.status(400).json({ message: "Token no valid" });
+
+        const challenges = await getConnection().get("detailChallenges").filter({ status: "doing" }).value();
+        console.log(challenges)
+        if (!challenges) return res.status(400).json({message: "No Challenges to show"});
+
+        res.status(200).json({ challenges: challenges });
+    } catch (error) {
+        res.status(500).json({ error, message: "There was a server error" });
+    }
+}
+
 exports.getAllChallenges = getAllChallenges;
 exports.takeChallenge = takeChallenge;
 exports.achieveChallenge = achieveChallenge;
 exports.getAchieved = getAchieved;
+exports.getCurrent = getCurrent;
